pref(sakura-web):人员信息平台字段类型调整

+++ sakura-web/src/main/java/com/sakura/common/web/domian/UserDTO.java
@@ -83,6 +83,6 @@ public class UserDTO {
      * 平台id
      */
     @ApiModelProperty(value = "平台id")
-    private Integer lqbPlatformId;
+    private Long lqbPlatformId;
 
 }
